Added ability to ban/unban players
Added new helper function to get amount of minutes that have passed since midnight 01/01/1970 UTC - GetMinutesSinceEpoch()
Converted wTimeBan property of CAccountBlock from UI16 to UI32
Added check during login for timeban account property. If a timeban is in place, player will be unable to login until after the ban is over.
Added JS property for Account JS Object:
	.timeban 			// Time in minutes that user is banned
Updated .get command to allow retrieving timeban value from account of player characters
Updated .set command to allow setting timeban value on account of player characters. If set to higher than 0, player will also be marked as banned, and disconnected
Added two new commands in js/commands/targeting/disconnect.js:
	'ban [durationInMinutes]	// Ban targeted player for specified duration in minutes, or for a default 24 hours if no duration is specified
	'unban 						// Unban targeted player
Updated 'tweak command (js/commands/targeting/tweak.js) to support viewing and modifying .timeban account property
Updated dictionary files with additional messages used in relation to banning/unbanning players

+++ source/cAccountClass.cpp
@@ -383,7 +383,7 @@ UI16 cAccountClass::CreateAccountSystem( void )
 		else if( l == "TIMEBAN" )
 		{
 			if( !r.empty() && r.length() != 0 )
-				actb.wTimeBan = r.toUShort();
+				actb.wTimeBan = r.toUInt();
 			else
 				actb.wTimeBan = 0;
 
@@ -1020,7 +1020,7 @@ UI16 cAccountClass::Load(void)
 		else if( l == "TIMEBAN" )
 		{
 			if( !r.empty() && r.length() != 0 )
-				actb.wTimeBan = UString( r ).toUShort();
+				actb.wTimeBan = UString( r ).toUInt();
 			else
 				actb.wTimeBan = 0;
